5 items seeded, a couple of bug fixes

+++ codebase/models/items.rb
@@ -14,8 +14,8 @@ def initialize(options)
   @quantity = options['quantity'].to_i
   @level = options['level'].to_i
   @effects = options['effects']
-  @buy_price = options['buy_price'].to_i
-  @sell_price = options['sell_price'].to_i
+  @buy_price = options['buy_price'].to_f
+  @sell_price = options['sell_price'].to_f
 end
